ゲーム進行周り完成

+++ EL0306/Assets/Tajima/Score.cs
@@ -1,5 +1,5 @@
 ﻿
 public static class Score
 {
-    public static float score = 0.0f;
+    public static int score = 0;
 }
